+++ src/Input/CustomAffix/CustomAffix.scss
