--- src/Input/Icon/index.js
@@ -1,1 +1,0 @@
-export { default } from './Icon.js';
