+++ src/Input/IconAffix/IconAffix.scss
@@ -6,7 +6,14 @@
     align-items: center;
     height: 100%;
     color: $B10;
-    &.pad-right {
+    &.inSuffix {
         padding-right: 6px;
+    }
+}
+
+@include withRTL() {
+    padding: 0;
+    &.inSuffix {
+        padding-left: 6px;
     }
 }
