+++ src/Input/IconAffix/index.js
@@ -1,0 +1,1 @@
+export { default } from './IconAffix.js';
